organized and added documentation
