--- CareerSystem/template/footer1.php
@@ -1,20 +1,0 @@
-			
-            <div class="cleaner h20"></div>
-            
-        	<div class="cleaner"></div>
-            
-        </div>
-        
-        <div id="tooplate_main_bottom"></div>
-                       
-    </div> <!-- end of tooplate_wrapper -->
-    
-    <div id="tooplate_footer_wrapper">
-        <div id="tooplate_footer">
-            <center><?php echo $content['sitename'].' &copy; '.date('Y') ?>. <?php echo isset( $_SESSION['loggedin_adminidame'] ) ? "Hello " .$_SESSION['loggedin_adminidame']."." : "" ?> All rights reserved.
-			</center>
-        </div>
-    </div>
-</div>
-</body>
-</html>
